requirements.txt
aws-cdk-lib==2.58.1
constructs>=10.0.0,<11.0.0
boto3>=1.26.41,<2.0.0
aws-lambda-powertools
requests
cdk-nag>=2.21.68